Add exeptions management , fix bugs update ,add dto perspective

+++ src/main/java/com/commerce/produit/service/business/ProductService.java
@@ -1,9 +1,13 @@
 package com.commerce.produit.service.business;
 
+import com.commerce.produit.service.exceptions.ResourceNotFoundException;
 import com.commerce.produit.service.model.Product;
 import com.commerce.produit.service.repository.ProductRepository;
+import com.sun.jdi.request.ExceptionRequest;
+import org.apache.logging.log4j.message.Message;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.xml.transform.ResourceSource;
 
 import java.util.List;
 import java.util.Optional;
@@ -26,11 +30,35 @@ public class ProductService {
         return productRepository.findById(id);
     }
 
-    public Product updateProduct(Product product) {
-        return productRepository.save(product);
-    }
+
+
+    public Product updateProduct(Product product, Long id) {
+
+        Optional<Product> existingProduct = productRepository.findById(id);
+        if (existingProduct.isPresent()) {
+           product.setId(id);
+            return productRepository.save(product);
+        } else {
+            String s = "Product not found with id " + product.getId();
+            throw new ResourceNotFoundException(s);
+        }
+
+        }
+
+
+
 
     public void deleteProduct(Long id) {
-        productRepository.deleteById(id);
+
+        Optional<Product> existingProduct = productRepository.findById(id);
+        if (existingProduct.isPresent()) {
+          productRepository.deleteById(id);;
+        } else {
+            String s = "Product not found with id " + id;
+            throw new ResourceNotFoundException(s);
+        }
+
     }
-}
+
+    }
+
